Move thread choosing logic to closure
Previously the callback was always called from the main thread blocking the UI
and potentially leading to unintended behavior. I came across this issue myself.

The suggested solution is to simply call the callback as is – if users want the
code to be run in the main thread they can still do this manually in their callback.
I've extended the method documentation to explain this.